Add nvim-client script
This is a wrapper around the `nvim` binary. When invoked outside the
context of a Neovim terminal, it just execs `nvim` with the given
arguments. When invoked inside a Neovim terminal, it opens a split
inside the existing Neovim instance and loads the given file(s) into
that split using `args`.